pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mosaic-tracker"
version = "0.1.0"
description = "Python implementation of MosaicSuite Particle Tracker for single particle tracking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Cell Cycle Laboratory", email = "lab@example.com"}
]
keywords = [
    "single particle tracking",
    "SPT",
    "particle tracking",
    "trajectory analysis",
    "MSD",
    "diffusion coefficient",
    "cell biology",
    "microscopy"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "pandas>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
]
viz = [
    "matplotlib>=3.5",
]
io = [
    "tifffile>=2021.0",
    "Pillow>=8.0",
]
all = [
    "matplotlib>=3.5",
    "tifffile>=2021.0",
    "Pillow>=8.0",
    "scikit-image>=0.19",
]

[project.urls]
"Homepage" = "https://github.com/your-lab/mosaic-tracker"
"Bug Tracker" = "https://github.com/your-lab/mosaic-tracker/issues"
"Documentation" = "https://github.com/your-lab/mosaic-tracker#readme"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=mosaic_tracker --cov-report=term-missing"
